Move tests to v1 scope
